# Switch benches to use `divan`