--- a/.github/workflows/ci.yaml
+++ b/.github/workflows/ci.yaml
@@ -39,7 +39,7 @@ jobs:
         rust:
           - stable
           - nightly
-          - 1.57
+          - "1.60"
         features:
           - default
           - ssl-openssl