--- a/.github/workflows/ci.yaml
+++ b/.github/workflows/ci.yaml
@@ -1,29 +1,17 @@
 on: [push, pull_request]
 name: CI
 jobs:
-  clippy_rustfmt:
-    name: Lint & Format
+  rustfmt:
+    name: Formatting
     runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        features:
-          - default
-          - ssl-openssl
-          - ssl-rustls
-          - ssl-native-tls
     steps:
       - uses: actions/checkout@v2
       - name: Install stable toolchain
         uses: actions-rs/toolchain@v1
         with:
           toolchain: stable
-          components: rustfmt, clippy
-
-      - name: Clippy
-        uses: actions-rs/cargo@v1
-        with:
-          command: clippy
-          args: --features ${{ matrix.features }}
+          override: true
+          components: rustfmt
 
       - name: Format
         uses: actions-rs/cargo@v1
@@ -31,8 +19,8 @@ jobs:
           command: fmt
           args: -- --check
 
-  test:
-    name: Build & Test
+  build_test_lint:
+    name: Build, Test, and Lints
     runs-on: ubuntu-latest
     strategy:
       matrix:
@@ -51,6 +39,7 @@ jobs:
         with:
           toolchain: ${{ matrix.rust }}
           override: true
+          components: clippy
 
       - name: Build
         uses: actions-rs/cargo@v1
@@ -64,7 +53,13 @@ jobs:
           command: test
           args: --features ${{ matrix.features }}
 
-  msrv_test:
+      - name: Clippy
+        uses: actions-rs/cargo@v1
+        with:
+          command: clippy
+          args: --features ${{ matrix.features }}
+
+  msrv_build:
     name: Build & Test on MSRV
     runs-on: ubuntu-latest
     steps: