--- a/.github/workflows/ci.yaml
+++ b/.github/workflows/ci.yaml
@@ -53,6 +53,12 @@ jobs:
           command: test
           args: --features ${{ matrix.features }}
 
+      - name: Test Benchmarks
+        uses: actions-rs/cargo@v1
+        with:
+          command: bench
+          args: --features ${{ matrix.features }} -- --test
+
       - name: Clippy
         uses: actions-rs/cargo@v1
         with: