--- a/Cargo.toml
+++ b/Cargo.toml
@@ -34,7 +34,12 @@ native-tls = { version = "0.2", optional = true }
 rustc-serialize = "0.3"
 sha1 = "0.6.0"
 fdlimit = "0.1"
+divan = { version = "0.1.16", default-features = false }
 
 [package.metadata.docs.rs]
 # Enable just one SSL implementation
 features = ["ssl-openssl"]
+
+[[bench]]
+name = "bench"
+harness = false