--- a/benches/bench.rs
+++ b/benches/bench.rs
@@ -1,12 +1,20 @@
-#![feature(test)]
-
 extern crate fdlimit;
-extern crate test;
 extern crate tiny_http;
 
+use divan::counter::ItemsCount;
+use divan::{bench, Bencher, Divan};
 use std::io::Write;
 use std::process::Command;
+use std::time::Duration;
 use tiny_http::Method;
+
+fn main() {
+    // Run registered benchmarks
+    Divan::default()
+        .min_time(Duration::from_secs(2))
+        .config_with_args()
+        .main();
+}
 
 #[test]
 #[ignore]
@@ -29,13 +37,13 @@ fn curl_bench() {
 }
 
 #[bench]
-fn sequential_requests(bencher: &mut test::Bencher) {
+fn sequential_requests(bencher: Bencher) {
     let server = tiny_http::Server::http("0.0.0.0:0").unwrap();
     let port = server.server_addr().to_ip().unwrap().port();
 
     let mut stream = std::net::TcpStream::connect(("127.0.0.1", port)).unwrap();
 
-    bencher.iter(|| {
+    bencher.bench_local(|| {
         (write!(stream, "GET / HTTP/1.1\r\nHost: localhost\r\n\r\n")).unwrap();
 
         let request = server.recv().unwrap();
@@ -47,13 +55,13 @@ fn sequential_requests(bencher: &mut test::Bencher) {
 }
 
 #[bench]
-fn parallel_requests(bencher: &mut test::Bencher) {
+fn parallel_requests(bencher: Bencher) {
     fdlimit::raise_fd_limit();
 
     let server = tiny_http::Server::http("0.0.0.0:0").unwrap();
     let port = server.server_addr().to_ip().unwrap().port();
 
-    bencher.iter(|| {
+    bencher.counter(ItemsCount::new(num_requests)).bench(|| {
         let mut streams = Vec::new();
 
         for _ in 0..1000usize {