_Builds on top of #274 to get CI passing_

TODO(cosmic): this PR bumps the MSRV up to 1.80 because the MSRV CI job is currently using `cargo test` instead of `cargo check`. Fix that at some point...

---

Switches from using the unstable `test` feature to using the [`divan` crate](https://crates.io/crates/divan) to run the benchmarks. The [`criterion` crate](https://crates.io/crates/criterion) is another popular alternative, but I prefer `divan` because it's a lot lighter in general

I also took the liberty of expanding the `parallel_requests` benchmark to run over a variety of number of requests instead of just 1000

### Current output

```console
$ cargo +nightly bench
... elided output
     Running benches/bench.rs (/home/wintermute/Programming/Repos/tiny-http/target/release/deps/bench-24b4c6b2c2a1c0fd)

running 3 tests
test curl_bench ... ignored
test parallel_requests   ... bench:  25,626,822.00 ns/iter (+/- 2,548,371.96)
test sequential_requests ... bench:      11,130.42 ns/iter (+/- 2,081.80)

test result: ok. 0 passed; 0 failed; 1 ignored; 2 measured; 0 filtered out; finished in 8.04s
```

### This PR

```console
$ cargo bench
... elided output
     Running benches/bench.rs (target/release/deps/bench-037f89c3c1b9dbe1)
Timer precision: 13 ns
bench                   fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ parallel_requests                  │               │               │               │         │
│  ├─ 10                141.3 µs      │ 782 µs        │ 224.3 µs      │ 235.6 µs      │ 8446    │ 8446
│  ├─ 100               2.104 ms      │ 25.01 ms      │ 2.377 ms      │ 2.519 ms      │ 795     │ 795
│  ╰─ 1000              23.88 ms      │ 53.79 ms      │ 26.41 ms      │ 27 ms         │ 100     │ 100
╰─ sequential_requests  9.731 µs      │ 416 ms        │ 11.07 µs      │ 54.42 µs      │ 38009   │ 38009
```